+++ ignite/pkg/cosmosclient/txservice_test.go
@@ -199,3 +199,111 @@ func TestTxServiceBroadcast(t *testing.T) {
 		})
 	}
 }
+
+func TestTxServiceBroadcastAsync(t *testing.T) {
+	var (
+		accountName = "bob"
+		passphrase  = "passphrase"
+		txHash      = []byte{1, 2, 3}
+		txHashStr   = hex.EncodeToString(txHash)
+	)
+	r, err := cosmosaccount.NewInMemory()
+	require.NoError(t, err)
+	a, _, err := r.Create(accountName)
+	require.NoError(t, err)
+	// Export created account to we can import it in the Client below.
+	key, err := r.Export(accountName, passphrase)
+	require.NoError(t, err)
+	sdkaddr, err := a.Record.GetAddress()
+	require.NoError(t, err)
+	msg := &banktypes.MsgSend{
+		FromAddress: sdkaddr.String(),
+		ToAddress:   "cosmos1k8e50d2d8xkdfw9c4et3m45llh69e7xzw6uzga",
+		Amount: sdktypes.NewCoins(
+			sdktypes.NewCoin("token", math.NewIntFromUint64(1)),
+		),
+	}
+	tests := []struct {
+		name             string
+		msg              sdktypes.Msg
+		opts             []cosmosclient.Option
+		expectedResponse *sdktypes.TxResponse
+		expectedError    string
+		setup            func(suite)
+	}{
+		{
+			name:          "fail: error not found",
+			msg:           msg,
+			expectedError: "make sure that your account has enough balance",
+
+			setup: func(s suite) {
+				s.expectPrepareFactory(sdkaddr)
+				s.signer.EXPECT().
+					Sign(mock.Anything, mock.Anything, "bob", mock.Anything, true).
+					Return(nil)
+				s.rpcClient.EXPECT().
+					BroadcastTxSync(mock.Anything, mock.Anything).
+					Return(nil, sdkerrors.ErrNotFound)
+			},
+		},
+		{
+			name:          "fail: response code > 0",
+			msg:           msg,
+			expectedError: "error code: '42' msg: 'oups'",
+
+			setup: func(s suite) {
+				s.expectPrepareFactory(sdkaddr)
+				s.signer.EXPECT().
+					Sign(mock.Anything, mock.Anything, "bob", mock.Anything, true).
+					Return(nil)
+				s.rpcClient.EXPECT().
+					BroadcastTxSync(mock.Anything, mock.Anything).
+					Return(&ctypes.ResultBroadcastTx{
+						Code: 42,
+						Log:  "oups",
+					}, nil)
+			},
+		},
+		{
+			name: "ok: tx broadcasted successfully",
+			msg:  msg,
+			expectedResponse: &sdktypes.TxResponse{
+				TxHash: txHashStr,
+				RawLog: "",
+			},
+
+			setup: func(s suite) {
+				s.expectPrepareFactory(sdkaddr)
+				s.signer.EXPECT().
+					Sign(mock.Anything, mock.Anything, "bob", mock.Anything, true).
+					Return(nil)
+				s.rpcClient.EXPECT().
+					BroadcastTxSync(mock.Anything, mock.Anything).
+					Return(&ctypes.ResultBroadcastTx{
+						Hash: txHash,
+					}, nil)
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			c := newClient(t, tt.setup, tt.opts...)
+			account, err := c.AccountRegistry.Import(accountName, key, passphrase)
+			require.NoError(t, err)
+			ctx := c.Context().
+				WithFromName(accountName).
+				WithFromAddress(sdkaddr)
+			txService, err := c.CreateTx(ctx.CmdContext, account, tt.msg)
+			require.NoError(t, err)
+
+			res, err := txService.BroadcastAsync(ctx.CmdContext)
+			if tt.expectedError != "" {
+				require.EqualError(t, err, tt.expectedError)
+				return
+			}
+			require.NoError(t, err)
+			require.Equal(t, ctx.Codec, res.Codec)
+			require.Equal(t, tt.expectedResponse, res.TxResponse)
+		})
+	}
+}